Separate fragment generation from MailMergeView
