+++ front/src/components/MailMerge/MailMerge.tsx
@@ -2,7 +2,6 @@ import React, { useState } from 'react';
 import SchemaSelector, { SchemaType } from './SchemaSelector';
 import View, { Props as ViewProps } from './MailMergeView';
 import Editor from './MailMergeEditor';
-import useUrlParams from 'utils/UrlParamsProvider';
 
 // MailMerge props includes all ViewProps except context is renamed to sample
 interface Props extends Omit<ViewProps, 'context'> {
@@ -27,8 +26,6 @@ export default function MailMerge(props: Props) {
     props.onTemplateChanged(before + templateString + after);
   };
   const style = { ...defaultStyle, ...props.style };
-  const urlParams = useUrlParams()
-  const queryStringParams = new Set(Object.keys(urlParams || {}));
   return (
     <>
       <div style={style}>
@@ -43,9 +40,6 @@ export default function MailMerge(props: Props) {
         style={{border: '2px solid black'}}
         template={template}
         context={props.sample}
-        fragmentName={props.fragmentName}
-        fragmentClass={props.fragmentClass}
-        onFragmentChanged={props.onFragmentChanged}
         islands={props.islands}
       />
     </>
