+++ front/src/components/MailMerge/index.ts
@@ -1,4 +1,5 @@
 import View from './MailMergeView';
 import MailMerge from './MailMerge';
+import { compileFragment } from './MailMergeFragment'
 
-export { MailMerge as MailMergeEditor, View as MailMergeView };
+export { MailMerge as MailMergeEditor, View as MailMergeView, compileFragment };
