+++ front/src/containers/GenericPage/GenericPage.tsx
@@ -12,6 +12,7 @@ import { studyIslands } from 'containers/Islands/CommonIslands'
 import useUrlParams from 'utils/UrlParamsProvider';
 import { find, propEq } from 'ramda';
 import {usePresentSite} from "../PresentSiteProvider/PresentSiteProvider";
+import { useFragment } from 'components/MailMerge/MailMergeFragment';
 
 
 interface Props {
@@ -37,13 +38,12 @@ export default function GenericPage(props: Props) {
     }
   }
   // When we add more page types we need to refactor this a little bit and pull out the query/nctid
-  const fragmentName = 'GenericPageStudy';
   const params = useUrlParams();
   const { site } = usePresentSite({ url: params.sv});
   const { data: pageViewsData } = usePageViews(site?.id);
   const { data: pageViewData } = usePageView(defaultPage());
   const currentPage = pageViewData?.site?.pageView;
-  const [fragment, setFragment] = useState('');
+  const [ fragmentName, fragment ] = useFragment('Study', currentPage?.template || '');
   const { data: studyData, loading } = useQuery(
     getStudyQuery(fragmentName, fragment),
     {
@@ -68,9 +68,6 @@ export default function GenericPage(props: Props) {
       <MailMergeView
         template={currentPage?.template || ''}
         context={studyData?.study}
-        fragmentName={fragmentName}
-        fragmentClass="Study"
-        onFragmentChanged={setFragment}
         islands={studyIslands}
       />
     </div>
